For the first exercise i had copied and pasted various ways to read different csv files with dung beetle data. the only table which seemed to be tidy was (4). This was because the others had not combined data into set variables. E.g. months were not combined, so Jan and Feb were put into separate columns
Unique() deletes the duplicate values or rows present in the vector, data frame, or matrix.
Length() finds out how many items are present in the vector

We are unable to use unique() for the beetles1 dataset so subscripting is used. It is the 3rd column to the end of the columns and prints out all the names of the columns.