read.table is used to read larger tables instead of using read.csv. this is what read.csv uses under the hood. you are able to change the default arrangements in files in many different formats.
in the cases for these files, they are .txt instead of .csv. if you use the same command of sep= " " like in .csv it won't put the data into a table format. sep="\t" is supposed to be used for .txt.
for the first .txt the only change that needs to be done is this.

for the second one there is an additional row which makes it a bit confusing, so in order to ignore it "skip" is used. "skip = 1" means skip the first row.

library(tidyr) downloads the tidy function that has been made by someone else. in this file the site number is only written once, so in order to have the site number written more i downloaded this function. it will also get confusing for the computer to only have each sit number printed only once, so it is important to fill in the empty spaces.
"?fill" searches what it means in R search bar

for data i put down beetlesdf (the file i wanted) and just wrote down "site" which is the column i wanted to fill.